@@ -7,9 +7,11 @@
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.1.0",
     "react-scripts": "5.0.0",
+    "react-transition-group": "^4.4.2",
     "sass": "^1.49.9",
     "web-vitals": "^2.1.4"
   },
