@@ -1,42 +1,33 @@
-import {useState} from "react";
-import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
+import {lazy, Suspense} from "react";
+import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
 
-import decoration from '../../resources/img/vision.png';
-import ComicsList from "../comicsList/ComicsList";
+import AppHeader from "../appHeader/AppHeader";
+import Spinner from "../spinner/Spinner";
+
+const Page404 = lazy(() => import('../pages/404'));
+const MainPage = lazy(() => import('../pages/MainPage'));
+const ComicsPage = lazy(() => import('../pages/ComicsPage'));
+const SingleComicPage = lazy(() => import('../pages/singleComicPage'));
 
 const App = () => {
 
-    const [selectedChar, setChar] = useState(null);
 
-   const onCharSelected = (id) => {
-        setChar(id);
-    }
     return (
-            <div className="app">
-                <AppHeader/>
-                <main>
-                    <ErrorBoundary>
-                        <ComicsList />
-                    </ErrorBoundary>
-
-                    {/*<ErrorBoundary>*/}
-                    {/*    <RandomChar/>*/}
-                    {/*</ErrorBoundary>*/}
-                    {/*<div className="char__content">*/}
-                    {/*    <ErrorBoundary>*/}
-                    {/*        <CharList onCharSelected={onCharSelected}/>*/}
-                    {/*    </ErrorBoundary>*/}
-                    {/*    <ErrorBoundary>*/}
-                    {/*        <CharInfo charId={selectedChar}/>*/}
-                    {/*    </ErrorBoundary>*/}
-                    {/*</div>*/}
-                    {/*<img className="bg-decoration" src={decoration} alt="vision"/>*/}
-                </main>
-            </div>
+            <Router>
+                <div className="app">
+                    <AppHeader />
+                    <main>
+                        <Suspense fallback={<Spinner />}>
+                            <Routes>
+                                <Route path="/" element={<MainPage />} />
+                                <Route path="/comics" element={<ComicsPage />} />
+                                <Route path="/comics/:comicId" element={<SingleComicPage />} />
+                                <Route path="*" element={<Page404 />} />
+                            </Routes>
+                        </Suspense>
+                    </main>
+                </div>
+            </Router>
         )
 
 }
