@@ -1,4 +1,7 @@
 import {useState, useEffect} from "react";
+import {Link} from "react-router-dom";
+//import {CSSTransition} from "react-transition-group";
+
 import PropTypes from "prop-types";
 import useMarvelService from "../../services/MarvelService";
 import Spinner from "../spinner/Spinner";
@@ -10,6 +13,7 @@ import './charInfo.scss';
 const CharInfo = (props) => {
 
     const [char, setChar] = useState(null);
+    const [inProp, setInProp] = useState(false);
 
     const {loading, error, clearError, getCharacter} = useMarvelService();
 
@@ -22,12 +26,15 @@ const CharInfo = (props) => {
         if (!charId) {
             return;
         }
+
         clearError();
-       getCharacter(charId).then(onCharLoaded);
+        getCharacter(charId).then(onCharLoaded);
     }
 
     const onCharLoaded = (char) => {
         setChar(char);
+        setInProp(true);
+        console.log(inProp);
     }
 
         const skeleton =  char || loading || error ? null : <Skeleton />
@@ -36,12 +43,17 @@ const CharInfo = (props) => {
         const content = !(loading || error || !char) ? <View char={char} /> : null;
 
         return (
-            <div className="char__info">
-                {skeleton}
-                {errorMessage}
-                {spinner}
-                {content}
-            </div>
+            <>
+                //<CSSTransition in={inProp} timeout={1500} classNames="char__info">
+                <div className="char__info">
+                    {skeleton}
+                    {errorMessage}
+                    {spinner}
+                    {content}
+                </div>
+                //</CSSTransition>
+            </>
+
         )
 }
 
@@ -52,39 +64,42 @@ const View = ({char}) => {
         imgStyle = {'objectFit': 'contain'}
     }
     return (
-        <>
-            <div className="char__basics">
-                <img src={thumbnail} alt={name} style={imgStyle}/>
-                <div>
-                    <div className="char__info-name">{name}</div>
-                    <div className="char__btns">
-                        <a href={homepage} className="button button__main">
-                            <div className="inner">homepage</div>
-                        </a>
-                        <a href={wiki} className="button button__secondary">
-                            <div className="inner">Wiki</div>
-                        </a>
+            <>
+                <div className="char__basics">
+                    <img src={thumbnail} alt={name} style={imgStyle}/>
+                    <div>
+                        <div className="char__info-name">{name}</div>
+                        <div className="char__btns">
+                            <a href={homepage} className="button button__main">
+                                <div className="inner">homepage</div>
+                            </a>
+                            <a href={wiki} className="button button__secondary">
+                                <div className="inner">Wiki</div>
+                            </a>
+                        </div>
                     </div>
                 </div>
-            </div>
-            <div className="char__descr">
-                {description}
-            </div>
-            <div className="char__comics">Comics:</div>
-            <ul className="char__comics-list">
-                {(comics.length > 0) ? null : 'There is no comics'}
-                {
-                    comics.map((item, i)=> {
-                        if (i > 9) return;
-                        return (
-                            <li key={i} className="char__comics-item">
-                             {item.name}
-                            </li>
-                        )
-                    })
-                }
-            </ul>
-        </>
+                <div className="char__descr">
+                    {description}
+                </div>
+                <div className="char__comics">Comics:</div>
+                <ul className="char__comics-list">
+                    {(comics.length > 0) ? null : 'There is no comics'}
+                    {
+                        comics.map((item, i)=> {
+                            if (i > 9) return;
+                            let url = item.resourceURI.slice(43, 48);
+                            return (
+                                <li key={i} className="char__comics-item">
+                                    <Link to={`/comics/${url}`}>
+                                        {item.name}
+                                    </Link>
+                                </li>
+                            )
+                        })
+                    }
+                </ul>
+            </>
     )
 }
 CharInfo.propTypes = {
