@@ -1,4 +1,5 @@
 import {useState, useEffect} from "react";
+
 import mjolnir from '../../resources/img/mjolnir.png';
 import useMarvelService from "../../services/MarvelService";
 import Spinner from "../spinner/Spinner";
@@ -59,15 +60,19 @@ const RandomChar = () => {
         )
 }
 
-const View = ({char}) => {
-    const {name, description, thumbnail, homepage, wiki} = char;
+const View = ({char}, props) => {
+    const {name, description, thumbnail, homepage, wiki, id} = char;
     let imgStyle = {'objectFit': 'cover'};
     if (thumbnail === 'http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg') {
         imgStyle = {'objectFit': 'contain'}
     }
     return (
         <div className="randomchar__block">
-            <img src={thumbnail} alt="Random character" className="randomchar__img" style={imgStyle}/>
+            <img src={thumbnail}
+                 alt="Random character"
+                 className="randomchar__img"
+                 style={imgStyle}
+            />
             <div className="randomchar__info">
                 <p className="randomchar__name">{name}</p>
                 <p className="randomchar__descr">{description}</p>
